Added Identity server and authentication in the AuctionService

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -7,6 +7,7 @@ using Data;
 using Dtos;
 using Entities;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -17,13 +18,16 @@ public class AuctionsController : ControllerBase
     private readonly AuctionDbContext _context;
     private readonly IMapper _mapper;
     private readonly IPublishEndpoint _publishEndpoint;
+    private readonly ILogger<AuctionsController> _logger;
 
 
-    public AuctionsController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint)
+    public AuctionsController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint,
+        ILogger<AuctionsController> logger)
     {
         _context = context;
         _mapper = mapper;
         _publishEndpoint = publishEndpoint;
+        _logger = logger;
     }
 
     [HttpGet]
@@ -51,21 +55,22 @@ public class AuctionsController : ControllerBase
 
         return _mapper.Map<AuctionDto>(auction);
     }
-
+    
     [HttpPost]
+    [Authorize]
     public async Task<ActionResult<AuctionDto>> CreateAuction([FromBody] CreateAuctionDto request, CancellationToken cancellationToken = default)
     {
         var auction = _mapper.Map<Auction>(request);
-        //TODO: add current user as seller;
-        auction.Seller = "system";
+
+
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
         
         var createdAuctionDto = _mapper.Map<AuctionDto>(auction);
 
-        Console.ForegroundColor = ConsoleColor.Green;
-        Console.WriteLine($"--> Publishing {nameof(AuctionCreated)} event with auction id: {auction.Id}");
-        
+        _logger.LogInformation("--> Publishing {Event} event with auction id: {AuctionId}", nameof(AuctionCreated), auction.Id);
+
         var auctionCreated = _mapper.Map<AuctionCreated>(createdAuctionDto);
         await _publishEndpoint.Publish(auctionCreated, cancellationToken);
 
@@ -73,13 +78,14 @@ public class AuctionsController : ControllerBase
 
         if (!result)
         {
-            Console.ForegroundColor = ConsoleColor.Red;
+            _logger.LogError("Auction was not created. Error while trying to save Auction to database");
             return BadRequest("Auction was not created. Please contact admin.");
         }
 
         return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, createdAuctionDto);
     }
     
+    [Authorize]
     [HttpPut("{id:guid}")]
     public async Task<ActionResult> UpdateAuction([FromRoute] Guid id, [FromBody] UpdateAuctionDto request, CancellationToken cancellationToken = default)
     {
@@ -89,8 +95,9 @@ public class AuctionsController : ControllerBase
 
         if (auction is null)
             return NotFound();
-        
-        //TODO: check seller to be current user
+
+        if (auction.Seller != User.Identity.Name)
+            return Forbid();
 
         auction.Item.Make = request.Make ?? auction.Item.Make;
         auction.Item.Model = request.Model ?? auction.Item.Model;
@@ -101,16 +108,22 @@ public class AuctionsController : ControllerBase
         auction.UpdatedAt = DateTime.UtcNow;
 
         var auctionUpdated = _mapper.Map<AuctionUpdated>(auction);
+        _logger.LogInformation("--> Publishing {Event} event with auction id: {AuctionId}", nameof(AuctionUpdated), auction.Id);
+        
         await _publishEndpoint.Publish(auctionUpdated, cancellationToken);
 
         var result = await _context.SaveChangesAsync(cancellationToken) > 0;
         
         if (!result)
+        {
+            _logger.LogError("Auction was not updated. Error while trying to update Auction to database");
             return BadRequest("Auction was not updated. Please contact admin.");
+        }
         
         return NoContent();
     }
 
+    [Authorize]
     [HttpDelete("{id:guid}")]
     public async Task<ActionResult> DeleteAuction([FromRoute] Guid id, CancellationToken cancellationToken = default)
     {
@@ -119,18 +132,24 @@ public class AuctionsController : ControllerBase
         if (auction is null)
             return NotFound();
 
-        // TODO: check seller == current user name
+        if (auction.Seller != User.Identity.Name)
+            return Forbid();
         
         _context.Items.Remove(auction.Item);
         _context.Auctions.Remove(auction);
 
         var auctionDeleted = _mapper.Map<AuctionDeleted>(auction);
+        
+        _logger.LogInformation("--> Publishing {Event} event with auction id: {AuctionId}", nameof(AuctionDeleted), auction.Id);
         await _publishEndpoint.Publish(auctionDeleted, cancellationToken);
         
         var result = await _context.SaveChangesAsync(cancellationToken) > 0;
         
         if (!result)
-            return BadRequest("Auction could not be deleted. Please contact admin.");
+        {
+            _logger.LogError("Auction was not deleted. Error while trying to delete Auction in database");
+            return BadRequest("Auction was not deleted. Please contact admin.");
+        }
         
         return Ok();
     }
